+++ R/json.R
@@ -1,0 +1,10 @@
+
+#' @importFrom jsonlite fromJSON
+#' @export
+jsonlite::fromJSON
+
+#' @importFrom jsonlite read_json
+#' @export
+jsonlite::read_json
+
+