+++ NEWS.md
@@ -1,4 +1,12 @@
 
+# GiottoUtils 0.1.12 (2024/09/27)
+
+## New
+- `pbar()` and `with_pbar()` as progress reporting utils. These are wrappers around *progressr* functionalities.
+- reexports of `fromJSON()` and `read_json()` from `jsonlite`.
+- `melt_matrix()` to emulate `reshape2::melt.matrix()` (deprecated package) for most use cases in Giotto.
+
+
 # GiottoUtils 0.1.11 (2024/08/22)
 
 ## New